rename make-live to make-draft-live
